Basic Set up for SQLite (Python Database) and MySQL (Cloud Database)
I've added basic files for what we may use, if we're going python all the way, we'd use SQLite, as it makes the database built into the system. Then we can just use more python or other languages for the frontend.

I've also got a file showing mySql setup, this lets the database be on a server that we can all connect to if we get a server for it, I think we'll be goin gfor SQLite but we can easily change out the backend